refactor upgrade command
